django/belt_reviewer/requirements.txt
appdirs==1.4.0
bcrypt==3.1.3
cffi==1.9.1
Django==1.10.5
django-debug-toolbar==1.7
django-extensions==1.7.6
olefile==0.44
packaging==16.8
pycparser==2.17
pyparsing==2.1.10
python-memcached==1.58
six==1.10.0
sqlparse==0.2.3

